svm/requirements.txt
# conda create -n svm python=3.8
libsvm-official==3.32.0 
numpy==1.24.4
scipy==1.10.1
icecream==2.1.3
cvxopt==1.3.2